Another POC with only two "real" files changed: the script runner in tools/ and .clang-tidy-rjl. you can run it multiple times with changed clang tidys.
This results in a tree that builds (yay!) but makes carnage of preprocessor
indention because we're seemingly running into 7 year old bugs in freaking
clang-format.

https://github.com/llvm/llvm-project/issues/47997

In another window, I have a message how the hanging indent for
access modifiers is now in (yay!) but half-tabs for case just isn't
a thing...but the upload of THAT crashes git.

The good news is that this sidesteps all that horror I wrote
about in the idempotence and recursive issues in our headers.

It's been an extremely long day (still waiting that Sandman from
yesterday) so I'm just going to throw this up for discussion.